+++ src/app/api/elasticsearch/search/combined-search/route.ts
@@ -1,184 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { NextRequest, NextResponse } from "next/server";
-import client from "@/utils-ts/elasticsearch";
 import { generateEmbedding } from "@/utils-ts/generateEmbeddings";
-
-interface SearchFilters {
-  uploadDate?: string;
-  location?: string;
-  tags?: string[];
-  vidID?: string; // Added this
-}
-
-interface SearchOptions {
-  filters?: SearchFilters;
-  from?: number;
-  size?: number;
-}
-
-interface ElasticsearchQuery {
-  index: string;
-  body: {
-    query: {
-      bool: {
-        should: any[];
-        filter: any[];
-        minimum_should_match?: number;
-      };
-    };
-    from: number;
-    size: number;
-    _source?:
-      | {
-          excludes?: string[];
-        }
-      | undefined;
-  };
-}
-
-interface VideoDocument {
-  vidTitle: string;
-  vidDescription: string;
-  tags: string[];
-  location?: string;
-  transcript?: string;
-  englishTranslation?: string;
-  transcriptEmbedding?: number[];
-  vidID: string; // Added this
-}
-
-interface SnippetDocument {
-  transcriptSnippet: string;
-  englishTranslation?: string;
-  snippetEmbedding?: number[];
-  vidID: string; // Added this
-}
-
-interface ElasticsearchHit {
-  _source: VideoDocument | SnippetDocument;
-  _score: number;
-}
-
-interface ElasticsearchResponse {
-  hits: {
-    hits: ElasticsearchHit[];
-    total: number;
-    max_score: number;
-  };
-}
-
-export async function combinedSearch(
-  keywords: string[],
-  queryEmbedding: number[],
-  index: "videos" | "video_snippets",
-  { filters = {}, from = 0, size = 10 }: SearchOptions,
-) {
-  try {
-    console.log("\n=== Combined Search Debug ===");
-    console.log("Input params:", {
-      keywords,
-      index,
-      filters,
-      hasEmbedding: !!queryEmbedding,
-      embeddingLength: queryEmbedding?.length,
-    });
-
-    const { uploadDate, location, tags, vidID } = filters;
-    const embeddingField =
-      index === "videos" ? "transcriptEmbedding" : "snippetEmbedding";
-
-    const combinedQuery: ElasticsearchQuery = {
-      index,
-      body: {
-        query: {
-          bool: {
-            should: [
-              {
-                multi_match: {
-                  query: keywords.join(" "),
-                  fields:
-                    index === "videos"
-                      ? [
-                          "vidTitle^3",
-                          "vidDescription^2",
-                          "tags^1.5",
-                          "location",
-                          "transcript",
-                          "englishTranslation",
-                        ]
-                      : ["transcriptSnippet", "englishTranslation"],
-                  fuzziness: "AUTO",
-                  type: "best_fields",
-                  boost: 0.2,
-                },
-              },
-              {
-                script_score: {
-                  query: { match_all: {} },
-                  script: {
-                    source: `
-                      double similarity = cosineSimilarity(params.query_vector, '${embeddingField}');
-                      double normalizedScore = 1.0 / (1.0 + Math.exp(-12.0 * (similarity - 0.55)));
-                      return normalizedScore * 0.8;
-                    `,
-                    params: { query_vector: queryEmbedding },
-                  },
-                },
-              },
-            ],
-            filter: [],
-            minimum_should_match: 1,
-          },
-        },
-        from,
-        size,
-        _source:
-          index === "videos"
-            ? {
-                excludes: [
-                  "transcriptJson.alternatives.words",
-                  "englishTranscriptJson.alternatives.words",
-                ],
-              }
-            : undefined,
-      },
-    };
-
-    const filterArray: any[] = [];
-
-    if (uploadDate && /^\d{4}-\d{2}-\d{2}$/.test(uploadDate)) {
-      filterArray.push({ term: { uploadDate } });
-    }
-    if (location) {
-      filterArray.push({ match: { location } });
-    }
-    if (tags && tags.length > 0) {
-      filterArray.push({ terms: { tags } });
-    }
-    if (vidID) {
-      filterArray.push({ term: { vidID } }); // Added this - works for both videos and snippets
-    }
-
-    combinedQuery.body.query.bool.filter = filterArray;
-
-    console.log(
-      "Executing combined search with query:",
-      JSON.stringify(combinedQuery, null, 2),
-    );
-
-    const response = (await client.search(
-      combinedQuery,
-    )) as ElasticsearchResponse;
-
-    return response.hits.hits.map((hit: ElasticsearchHit) => ({
-      score: hit._score,
-      ...hit._source,
-    }));
-  } catch (error) {
-    console.error("\nCombined search error:", error);
-    throw error;
-  }
-}
+import { combinedSearch } from "../functions";
+import { SearchFilters } from "@/app/api/interfaces";
 
 export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams;
