+++ src/app/api/elasticsearch/search/keyword-search/route.ts
@@ -1,195 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { NextRequest, NextResponse } from "next/server";
-import client from "@/utils-ts/elasticsearch";
-
-interface VideoDocument {
-  vidDescription: string;
-  tags: string[];
-  location?: string;
-  transcript?: string;
-  englishTranslation?: string;
-  vidID: string;
-}
-
-interface SnippetDocument {
-  transcriptSnippet: string;
-  englishTranslation?: string;
-  vidID: string;
-}
-
-interface SearchFilters {
-  uploadDate?: string;
-  location?: string;
-  tags?: string[];
-  vidID?: string;
-}
-
-interface SearchParams {
-  keywords: string | string[];
-  filters?: SearchFilters;
-  from?: number;
-  size?: number;
-}
-
-interface ElasticsearchQuery {
-  index: string;
-  body: {
-    query: {
-      bool: {
-        must: any[];
-        filter: any[];
-      };
-      multi_match?: {
-        query: string;
-        fields: string[];
-        type: string;
-        fuzziness: string;
-      };
-    };
-    from: number;
-    size: number;
-    _source?: {
-      excludes?: string[];
-    };
-  };
-}
-
-interface ElasticsearchHit {
-  _source: VideoDocument | SnippetDocument;
-  _score: number;
-}
-
-interface ElasticsearchResponse {
-  hits: {
-    hits: ElasticsearchHit[];
-    total: number;
-    max_score: number;
-  };
-}
-
-function determineSearchField(keywords: string | string[]): string {
-  const nepaliPattern = /[ऀ-ॿ]/;
-  const keywordString = Array.isArray(keywords) ? keywords.join(" ") : keywords;
-  return nepaliPattern.test(keywordString)
-    ? "transcript"
-    : "englishTranslation";
-}
-
-export async function searchVideos({
-  keywords,
-  filters = {},
-  from = 0,
-  size = 10,
-}: SearchParams) {
-  const { uploadDate, location, tags, vidID } = filters;
-  const searchField = determineSearchField(keywords);
-  const keywordString = Array.isArray(keywords) ? keywords.join(" ") : keywords;
-  console.log(`keywordString ->:`, keywordString);
-
-  const filterArray: any[] = [];
-
-  if (uploadDate && /^\d{4}-\d{2}-\d{2}$/.test(uploadDate)) {
-    filterArray.push({ term: { uploadDate } });
-  }
-  if (location) {
-    filterArray.push({ match: { location } });
-  }
-  if (tags && tags.length > 0) {
-    filterArray.push({ terms: { tags } });
-  }
-  if (vidID) {
-    filterArray.push({ term: { vidID } }); // Added this
-  }
-
-  const videoSearchQuery: ElasticsearchQuery = {
-    index: "videos",
-    body: {
-      query: {
-        bool: {
-          must: [
-            {
-              multi_match: {
-                query: keywordString,
-                fields: [
-                  "vidTitle^3",
-                  "vidDescription^2",
-                  "tags^1.5",
-                  searchField,
-                ],
-                type: "best_fields",
-                fuzziness: "AUTO",
-              },
-            },
-          ],
-          filter: filterArray,
-        },
-      },
-      from,
-      size,
-      _source: {
-        excludes: [
-          "transcriptJson.alternatives.words",
-          "englishTranscriptJson.alternatives.words",
-        ],
-      },
-    },
-  };
-
-  const response = (await client.search(
-    videoSearchQuery,
-  )) as ElasticsearchResponse;
-  return response.hits.hits.map((hit: ElasticsearchHit) => ({
-    ...hit._source,
-    score: hit._score,
-  }));
-}
-
-export async function searchSnippets({
-  keywords,
-  filters = {},
-  from = 0,
-  size = 10,
-}: SearchParams) {
-  const searchField = determineSearchField(keywords);
-  const keywordString = Array.isArray(keywords) ? keywords.join(" ") : keywords;
-  const { vidID } = filters;
-
-  const filterArray: any[] = [];
-  if (vidID) {
-    filterArray.push({ term: { vidID } });
-  }
-
-  const snippetSearchQuery: ElasticsearchQuery = {
-    index: "video_snippets",
-    body: {
-      query: {
-        bool: {
-          must: [
-            {
-              multi_match: {
-                query: keywordString,
-                fields: [searchField],
-                type: "best_fields",
-                fuzziness: "AUTO",
-              },
-            },
-          ],
-          filter: filterArray,
-        },
-      },
-      from,
-      size,
-    },
-  };
-
-  const response = (await client.search(
-    snippetSearchQuery,
-  )) as ElasticsearchResponse;
-  return response.hits.hits.map((hit: ElasticsearchHit) => ({
-    ...hit._source,
-    score: hit._score,
-  }));
-}
+import { SearchFilters } from "@/app/api/interfaces";
+import { searchSnippets, searchVideos } from "../functions";
 
 export async function GET(request: NextRequest) {
   const searchParams = request.nextUrl.searchParams;
